scale input weights
Before this change, the encoder's outputs all started with standard deviations on the order of 0.03. Now, after the change, we have stats more like this:

115 mean=0.140  stddev=0.399
116 mean=-0.076 stddev=0.398
117 mean=0.150  stddev=0.357
118 mean=0.231  stddev=0.429
119 mean=0.020  stddev=0.330

This is beneficial, because it means that much more usable information will enter the decoder from the encoder. Before this change, I observed that the decoder--especially a deep and large one--often ignored tons of components coming from the encoder. The result was that the encoder produced stats more like this:

1016    mean=0.994      stddev=0.001
1017    mean=0.977      stddev=0.017
1018    mean=-0.072     stddev=0.834
1019    mean=0.398      stddev=0.775
1020    mean=0.973      stddev=0.007
1021    mean=0.965      stddev=0.012
1022    mean=-0.986     stddev=0.006
1023    mean=-0.981     stddev=0.005